Modifica UC1 con aggiunta classe Catalogo
