+++ src/main/java/com/admiral/App.java
@@ -4,9 +4,7 @@ public class App {
 
     public static void main(String[] args) {
         Admiral admiral = Admiral.getInstance();
-        
-        admiral.loadDestinazioni();
-        admiral.loadPorti();
-        admiral.loadNavi();
+
+        System.out.println("Applicazione avviata e Catalogo inizializzato.");
     }
 }
