+++ src/main/java/com/admiral/Itinerario.java
@@ -15,49 +15,49 @@ public class Itinerario {
     private Nave nave;
     private Porto portoPartenza;
     private Map<String, Porto> portiDaVisitare;
-    
+
     public Itinerario(String codice, String codiceDestinazione, String codiceNave, String codicePortoPartenza,
-                        LocalDate dataPartenza, LocalDate dataRitorno){
-        Admiral admiral = Admiral.getInstance();
+            LocalDate dataPartenza, LocalDate dataRitorno) {
+        Catalogo catalogo = Catalogo.getInstance();
 
         this.codice = codice;
-        this.destinazione = admiral.getDestinazione(codiceDestinazione);
-        this.nave = admiral.getNave(codiceNave);
-        this.portoPartenza = admiral.getPorto(codicePortoPartenza);
+        this.destinazione = catalogo.getDestinazione(codiceDestinazione);
+        this.nave = catalogo.getNave(codiceNave);
+        this.portoPartenza = catalogo.getPorto(codicePortoPartenza);
         this.dataPartenza = dataPartenza;
         this.dataRitorno = dataRitorno;
 
         this.portiDaVisitare = new HashMap<>();
     }
 
-    public void inserisciPortoDaVisitare(String codicePorto, Porto p){
+    public void inserisciPortoDaVisitare(String codicePorto, Porto p) {
         this.portiDaVisitare.put(codicePorto, p);
     }
 
-    public String getCodice(){
+    public String getCodice() {
         return codice;
     }
 
-    public String toString(){
+    public String toString() {
         String s = "Itinerario " + codice + "\n"
-                    + "Destinazione: " + destinazione + "\n"
-                    + "Porto di partenza: " + portoPartenza +"\n"
-                    + "Nave: " + nave + "\n"
-                    + "Prezzo per persona: " + destinazione.getPrezzo() + "\n"
-                    + "Data di partenza: " + dataPartenza + "  Data di ritorno: " + dataRitorno + "\n"
-                    + "Durata: " + durata + " giorni\n"
-                    + "Lista Porti da visitare: \n";
+                + "Destinazione: " + destinazione + "\n"
+                + "Porto di partenza: " + portoPartenza + "\n"
+                + "Nave: " + nave + "\n"
+                + "Prezzo per persona: " + destinazione.getPrezzo() + "\n"
+                + "Data di partenza: " + dataPartenza + "  Data di ritorno: " + dataRitorno + "\n"
+                + "Durata: " + durata + " giorni\n"
+                + "Lista Porti da visitare: \n";
         for (Porto porto : portiDaVisitare.values()) {
-                    s += "\t" + porto.getNome() + "\n";    
+            s += "\t" + porto.getNome() + "\n";
         }
         return s;
     }
 
-    public Map<String, Porto> getPortiDaVisitare(){
+    public Map<String, Porto> getPortiDaVisitare() {
         return this.portiDaVisitare;
     }
 
-    public Porto getPorto(String codice){
+    public Porto getPorto(String codice) {
         return portiDaVisitare.get(codice);
     }
 }
